ALFMOB-22: Integrated dropdown for size selection with more than 6 items
